+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.8...@kong-ui-public/analytics-config-store@1.4.9) (2026-08-03)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
+## [1.4.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.7...@kong-ui-public/analytics-config-store@1.4.8) (2026-07-30)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [1.4.7](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@1.4.6...@kong-ui-public/analytics-config-store@1.4.7) (2026-07-29)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store