+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.6.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.6.10...@kong-ui-public/analytics-metric-provider@11.6.11) (2026-08-03)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.6.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.6.9...@kong-ui-public/analytics-metric-provider@11.6.10) (2026-07-30)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.6.9](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.6.8...@kong-ui-public/analytics-metric-provider@11.6.9) (2026-07-29)