+++ packages/entities/entities-routes/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.26.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.26.2...@kong-ui-public/entities-routes@3.26.3) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.26.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.26.1...@kong-ui-public/entities-routes@3.26.2) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+## [3.26.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.26.0...@kong-ui-public/entities-routes@3.26.1) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-routes
+
+
+
+
+
+# [3.26.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.25.8...@kong-ui-public/entities-routes@3.26.0) (2026-07-30)
+
+
+### Features
+
+* **expressions:** migrate editor to @kong-ui-public/monaco-editor ([#3620](https://github.com/Kong/public-ui-components/issues/3620)) ([82e4e3a](https://github.com/Kong/public-ui-components/commit/82e4e3a6590adacf790cb6fd0b3de967a2b61e77))
+
+
+
+
+
 ## [3.25.8](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-routes@3.25.7...@kong-ui-public/entities-routes@3.25.8) (2026-07-29)
 
 **Note:** Version bump only for package @kong-ui-public/entities-routes