+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.16.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.16.12...@kong-ui-public/entities-upstreams-targets@3.16.13) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [3.16.12](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.16.11...@kong-ui-public/entities-upstreams-targets@3.16.12) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [3.16.11](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.16.10...@kong-ui-public/entities-upstreams-targets@3.16.11) (2026-07-31)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [3.16.10](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.16.9...@kong-ui-public/entities-upstreams-targets@3.16.10) (2026-07-29)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets